refactor(global): moving delete question group state into global state

+++ components/books/question-group/delete-form.tsx
@@ -1,14 +1,10 @@
 "use client";
-import { useForm } from "react-hook-form";
-import { z } from "zod";
-import { zodResolver } from "@hookform/resolvers/zod";
 import { toast } from "sonner";
-import { useTransition } from "react";
+import { useEffect, useTransition } from "react";
 import { useRouter } from "next/navigation";
-import { QuestionGroup } from "@prisma/client";
-import { QuestionGroupSchema } from "@/lib/validations/books";
 import { deleteQuestionGroup } from "@/actions/books/question-group";
 import {
+  AlertDialog,
   AlertDialogAction,
   AlertDialogCancel,
   AlertDialogContent,
@@ -17,28 +13,23 @@ import {
   AlertDialogHeader,
   AlertDialogTitle,
 } from "@/components/ui/alert-dialog";
+import { useEditHook } from "@/global/use-edit-hook";
 
-export function DeleteQuestionGroupForm({
-  questionGroup,
-  setIsCUD,
-}: {
-  questionGroup: QuestionGroup;
-  setIsCUD: (isCUD: boolean) => void;
-}) {
+export function DeleteQuestionGroupForm() {
   const [isPending, startTransition] = useTransition();
-  const form = useForm<z.infer<typeof QuestionGroupSchema>>({
-    resolver: zodResolver(QuestionGroupSchema),
-    defaultValues: {
-      title: questionGroup.title,
-      type: questionGroup.type,
-      startQuestionNumber: questionGroup.startQuestionNumber,
-      endQuestionNumber: questionGroup.endQuestionNumber,
-      description: questionGroup.description || "",
-    },
-  });
+  const { data, isOpen, type, onClose } = useEditHook();
+  const isModalOpen = isOpen && type === "deleteQuestionGroup";
+  const questionGroup = data?.questionGroup;
+
   const router = useRouter();
 
+  if (!isModalOpen && !questionGroup) {
+    return null;
+  }
   const onSubmit = async () => {
+    if (!questionGroup) {
+      return;
+    }
     try {
       startTransition(async () => {
         const { success, error } = await deleteQuestionGroup({
@@ -46,7 +37,6 @@ export function DeleteQuestionGroupForm({
         });
 
         if (success) {
-          form.reset();
           router.refresh();
           toast.success("Successfully deleted questionGroup!");
         } else {
@@ -57,25 +47,27 @@ export function DeleteQuestionGroupForm({
       console.error("Error deleting questionGroup:", e);
       toast.error("Failed to delete questionGroup.");
     } finally {
-      setIsCUD(false);
+      onClose();
     }
   };
 
   return (
-    <AlertDialogContent>
-      <AlertDialogHeader>
-        <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
-        <AlertDialogDescription>
-          This action cannot be undone. This will permanently delete your
-          account and remove your data from our servers.
-        </AlertDialogDescription>
-      </AlertDialogHeader>
-      <AlertDialogFooter>
-        <AlertDialogCancel disabled={isPending}>Cancel</AlertDialogCancel>
-        <AlertDialogAction disabled={isPending} onClick={onSubmit}>
-          Continue
-        </AlertDialogAction>
-      </AlertDialogFooter>
-    </AlertDialogContent>
+    <AlertDialog open={isModalOpen} onOpenChange={onClose}>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+          <AlertDialogDescription>
+            This action cannot be undone. This will permanently delete your
+            account and remove your data from our servers.
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel disabled={isPending}>Cancel</AlertDialogCancel>
+          <AlertDialogAction disabled={isPending} onClick={onSubmit}>
+            Continue
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
   );
 }
